myproject/requirement.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.1
Deprecated==1.2.13
Django==4.2
django-cors-headers==3.14.0
django-oauth-toolkit==2.2.0
djangorestframework==3.14.0
idna==3.4
jwcrypto==1.4.2
oauthlib==3.2.2
pycparser==2.21
pytz==2023.3
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.15
wrapt==1.15.0
